backend/requirements.txt
fastapi>=0.110.0,<1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
